Make btns buttons, inherit font family
Normalize was resetting font family of buttons to a general sans-serif. This still has theoretical problems of inheriting fonts for specific sections, but I don’t know that it’s going to be a realistic issue for us.

Also makes the index page use `<button>` instead of the fictional `<btn>`